+++ app/src/mobile/java/com/ssmhdssmhd/mxboxs/ui/activity/LiveActivity.java
@@ -154,6 +154,7 @@ public class LiveActivity extends PlaybackActivity implements CustomKeyDown.List
 
     @Override
     protected void onServiceConnected() {
+        player().setLiveMode(true);
         PlaybackAction.setPlaybackMode(player(), mBinding.control.action.player, mBinding.control.action.decode);
         PlaybackAction.setSpeedText(player(), mBinding.control.action.speed);
         checkLive();