+++ app/src/mobile/java/com/ssmhdssmhd/mxboxs/ui/fragment/SettingPlayerFragment.java
@@ -68,6 +68,7 @@ public class SettingPlayerFragment extends BaseFragment implements UaListener, S
     @Override
     protected void initEvent() {
         mBinding.engine.setOnClickListener(this::setEngine);
+        mBinding.liveEngine.setOnClickListener(this::setLiveEngine);
         mBinding.mpvConf.setOnClickListener(this::onMpvConf);
         mBinding.mpvGpuNext.setOnClickListener(this::setMpvGpuNext);
         mBinding.mpvVulkan.setOnClickListener(this::setMpvVulkan);
@@ -113,6 +114,12 @@ public class SettingPlayerFragment extends BaseFragment implements UaListener, S
         setVisible();
     }
 
+    private void setLiveEngine(View view) {
+        int index = (PlayerSetting.getLiveEngine() + 1) % engine.length;
+        PlayerSetting.putLiveEngine(index);
+        setPlaybackModeText();
+    }
+
     private void onMpvConf(View view) {
         MpvConfDialog.show(this);
     }
@@ -137,6 +144,7 @@ public class SettingPlayerFragment extends BaseFragment implements UaListener, S
         engine = ResUtil.getStringArray(R.array.select_engine);
         render = ResUtil.getStringArray(R.array.select_render);
         mBinding.engineText.setText(engine[PlayerSetting.getEngine()]);
+        mBinding.liveEngineText.setText(engine[PlayerSetting.getLiveEngine()]);
         mBinding.renderText.setText(render[PlayerSetting.getRender()]);
     }